+++ src/Apply.ts
@@ -15,6 +15,7 @@
 import { flow, pipe } from './function'
 import { Functor, Functor1, Functor2, Functor2C, Functor3, Functor3C, Functor4 } from './Functor'
 import { HKT, Kind, Kind2, Kind3, Kind4, URIS, URIS2, URIS3, URIS4 } from './HKT'
+import { getDualSemigroup, Semigroup } from './Semigroup'
 
 /**
  * @category type classes
@@ -265,4 +266,33 @@ export function apT_<F>(
       F.map((a) => (b: B): readonly [...A, B] => [...a, b]),
       F.ap(fb)
     )
+}
+
+/**
+ * Lift a semigroup into 'F', the inner values are concatenated using the provided `Semigroup`.
+ *
+ * @since 3.0.0
+ */
+export function getApplySemigroup<F extends URIS4>(
+  F: Apply4<F>
+): <A, S, R, E>(S: Semigroup<A>) => Semigroup<Kind4<F, S, R, E, A>>
+export function getApplySemigroup<F extends URIS3>(
+  F: Apply3<F>
+): <A, R, E>(S: Semigroup<A>) => Semigroup<Kind3<F, R, E, A>>
+export function getApplySemigroup<F extends URIS3, E>(
+  F: Apply3C<F, E>
+): <A, R>(S: Semigroup<A>) => Semigroup<Kind3<F, R, E, A>>
+export function getApplySemigroup<F extends URIS2>(F: Apply2<F>): <A, E>(S: Semigroup<A>) => Semigroup<Kind2<F, E, A>>
+export function getApplySemigroup<F extends URIS2, E>(
+  F: Apply2C<F, E>
+): <A>(S: Semigroup<A>) => Semigroup<Kind2<F, E, A>>
+export function getApplySemigroup<F extends URIS>(F: Apply1<F>): <A>(S: Semigroup<A>) => Semigroup<Kind<F, A>>
+export function getApplySemigroup<F>(F: Apply<F>): <A>(S: Semigroup<A>) => Semigroup<HKT<F, A>>
+export function getApplySemigroup<F>(F: Apply<F>): <A>(S: Semigroup<A>) => Semigroup<HKT<F, A>> {
+  return <A>(S: Semigroup<A>) => {
+    const f = getDualSemigroup(S).concat
+    return {
+      concat: (second: HKT<F, A>) => (first: HKT<F, A>) => pipe(first, F.map(f), F.ap(second))
+    }
+  }
 }