+++ src/Either.ts
@@ -756,38 +756,6 @@ export const getSemigroup = <A, E>(S: Semigroup<A>): Semigroup<Either<E, A>> => 
 })
 
 /**
- * Semigroup returning the left-most `Left` value. If both operands are `Right`s then the inner values
- * are concatenated using the provided `Semigroup`
- *
- * @example
- * import * as E from 'fp-ts/Either'
- * import { semigroupSum } from 'fp-ts/Semigroup'
- * import { pipe } from 'fp-ts/function'
- *
- * const S = E.getApplySemigroup<number, string>(semigroupSum)
- * assert.deepStrictEqual(pipe(E.left('a'), S.concat(E.left('b'))), E.left('a'))
- * assert.deepStrictEqual(pipe(E.left('a'), S.concat(E.right(2))), E.left('a'))
- * assert.deepStrictEqual(pipe(E.right(1), S.concat(E.left('b'))), E.left('b'))
- * assert.deepStrictEqual(pipe(E.right(1), S.concat(E.right(2))), E.right(3))
- *
- * @category instances
- * @since 3.0.0
- */
-export const getApplySemigroup = <A, E>(S: Semigroup<A>): Semigroup<Either<E, A>> => ({
-  concat: (second) => (first) =>
-    isLeft(first) ? first : isLeft(second) ? second : right(S.concat(second.right)(first.right))
-})
-
-/**
- * @category instances
- * @since 3.0.0
- */
-export const getApplyMonoid = <A, E>(M: Monoid<A>): Monoid<Either<E, A>> => ({
-  concat: getApplySemigroup<A, E>(M).concat,
-  empty: right(M.empty)
-})
-
-/**
  * Builds a `Compactable` instance for `Either` given `Monoid` for the left side.
  *
  * @category instances