+++ test/TaskEither.ts
@@ -175,34 +175,12 @@ describe('TaskEither', () => {
     })
   })
 
-  describe('getSemigroup', () => {
-    it('concat', async () => {
-      const S = _.getSemigroup<number, string>(semigroupSum)
-      assert.deepStrictEqual(await pipe(_.left('a'), S.concat(_.left('b')))(), E.left('a'))
-      assert.deepStrictEqual(await pipe(_.left('a'), S.concat(_.right(2)))(), E.right(2))
-      assert.deepStrictEqual(await pipe(_.right(1), S.concat(_.left('b')))(), E.right(1))
-      assert.deepStrictEqual(await pipe(_.right(1), S.concat(_.right(2)))(), E.right(3))
-    })
-  })
-
-  describe('getApplyMonoid', () => {
-    const M = _.getApplyMonoid(monoidString)
-
-    it('concat (right)', async () => {
-      return assert.deepStrictEqual(await pipe(_.right('a'), M.concat(_.right('b')))(), E.right('ab'))
-    })
-
-    it('concat (left)', async () => {
-      return assert.deepStrictEqual(await pipe(_.right('a'), M.concat(_.left('b')))(), E.left('b'))
-    })
-
-    it('empty (right)', async () => {
-      return assert.deepStrictEqual(await pipe(_.right('a'), M.concat(M.empty))(), E.right('a'))
-    })
-
-    it('empty (left)', async () => {
-      return assert.deepStrictEqual(await pipe(M.empty, M.concat(_.right('a')))(), E.right('a'))
-    })
+  it('getSemigroup', async () => {
+    const S = _.getSemigroup<number, string>(semigroupSum)
+    assert.deepStrictEqual(await pipe(_.left('a'), S.concat(_.left('b')))(), E.left('a'))
+    assert.deepStrictEqual(await pipe(_.left('a'), S.concat(_.right(2)))(), E.right(2))
+    assert.deepStrictEqual(await pipe(_.right(1), S.concat(_.left('b')))(), E.right(1))
+    assert.deepStrictEqual(await pipe(_.right(1), S.concat(_.right(2)))(), E.right(3))
   })
 
   it('ApplicativeSeq', async () => {